fix: added more tests

+++ states/file.go
@@ -11,35 +11,32 @@ import (
 	"sync"
 )
 
+const FilePrefix = ".json"
+
 type FileStore[V any] struct {
-	mutex     sync.Mutex
-	path      string
-	mode      os.FileMode
-	_fullPath string
+	mutex    sync.Mutex
+	path     string
+	mode     os.FileMode
+	isClosed bool
 }
 
-func addSuffixToPath(path, suffix string) string {
-	ext := filepath.Ext(path)
-	base := strings.TrimSuffix(path, ext) // Remove original extension
-	return base + suffix + ext
-}
-
-func (sm *FileStore[V]) getFullPath(key string) string {
-	if len(sm._fullPath) == 0 {
-		sm._fullPath = addSuffixToPath(sm.path, "-"+strings.ReplaceAll(strings.ToLower(key), " ", "-"))
-	}
-	return sm._fullPath
+func (sm *FileStore[V]) getItemPath(key string) string {
+	return filepath.Join(sm.path, strings.ReplaceAll(strings.ToLower(key), " ", "-")+FilePrefix)
 }
 
 // Sets the value for a key.
 func (sm *FileStore[V]) Store(ctx *context.Context, key string, value V) error {
+	if sm.isClosed {
+		return ErrClosed
+	}
+
 	// To JSON
 	data, err := json.MarshalIndent(value, "", "  ")
 	if err != nil {
 		return fmt.Errorf("error marshaling state: Error: %w", err)
 	}
 	// Save to file
-	path := sm.getFullPath(key)
+	path := sm.getItemPath(key)
 	err = os.WriteFile(path, data, sm.mode)
 	if err != nil {
 		return fmt.Errorf("error saving state file: %s. Error: %w", path, err)
@@ -50,7 +47,11 @@ func (sm *FileStore[V]) Store(ctx *context.Context, key string, value V) error {
 // Load returns the value stored in the map for a key, or nil if no value is present.
 // The ok result indicates whether a value was found in the map.
 func (sm *FileStore[V]) Load(ctx *context.Context, key string) (value V, ok bool) {
-	path := sm.getFullPath(key)
+	if sm.isClosed {
+		return value, false
+	}
+
+	path := sm.getItemPath(key)
 
 	// Read from the file
 	data, err := os.ReadFile(path)
@@ -71,7 +72,11 @@ func (sm *FileStore[V]) Load(ctx *context.Context, key string) (value V, ok bool
 
 // Deletes the value for a key.
 func (sm *FileStore[V]) Delete(ctx *context.Context, key string) error {
-	path := sm.getFullPath(key)
+	if sm.isClosed {
+		return ErrClosed
+	}
+
+	path := sm.getItemPath(key)
 	err := os.Remove(path)
 	if err != nil && !os.IsNotExist(err) {
 		return fmt.Errorf("error deleting state file: %s. Error: %v", path, err)
@@ -81,23 +86,33 @@ func (sm *FileStore[V]) Delete(ctx *context.Context, key string) error {
 
 // Deletes the value for a key.
 func (sm *FileStore[V]) Close(ctx *context.Context) error {
+	sm.isClosed = true
 	return nil
 }
 
 // Deletes the value for a key.
 func (sm *FileStore[V]) Clear(ctx *context.Context) error {
-	return nil
+	if sm.isClosed {
+		return ErrClosed
+	}
+
+	return os.RemoveAll(sm.path)
 }
 
 // Creates and returns a new state store.
 func NewFileStateStore(basePath, storeName string) *FileStore[State] {
-	stateFilePath, err := filepath.Abs(filepath.Join(basePath, storeName))
+	path, err := filepath.Abs(filepath.Join(basePath, storeName))
 	if err != nil {
-		stateFilePath = storeName
+		path = filepath.Join(os.TempDir(), basePath, storeName)
 	}
+
+	if err := os.MkdirAll(path, 0644); err != nil {
+		panic(fmt.Sprintf("FATAL: could not create store directory: %v", err))
+	}
+
 	return &FileStore[State]{
 		mutex: sync.Mutex{},
-		path:  stateFilePath,
+		path:  path,
 		mode:  0644,
 	}
 }
